refactor(AddRoleDrawer): decompose into hooks and components
