+++ src/components/[guild]/AddAndOrderRoles/components/AddRoleDrawer.tsx
@@ -10,67 +10,24 @@ import {
   useDisclosure,
   VStack,
 } from "@chakra-ui/react"
-import useGuild from "components/[guild]/hooks/useGuild"
 import AddRolePlatforms from "components/[guild]/RolePlatforms/AddRolePlatforms"
-import useVisibilityModalProps from "components/[guild]/SetVisibility/hooks/useVisibilityModalProps"
 import Button from "components/common/Button"
 import DiscardAlert from "components/common/DiscardAlert"
-import DrawerHeader from "components/common/DrawerHeader"
 import Section from "components/common/Section"
 import Description from "components/create-guild/Description"
 import DynamicDevTool from "components/create-guild/DynamicDevTool"
-import useCreateRole, {
-  RoleToCreate,
-} from "components/create-guild/hooks/useCreateRole"
 import IconSelector from "components/create-guild/IconSelector"
 import Name from "components/create-guild/Name"
 import SetRequirements from "components/create-guild/Requirements"
-import usePinata from "hooks/usePinata"
-import useSubmitWithUpload from "hooks/useSubmitWithUpload"
-import useToast from "hooks/useToast"
 import useWarnIfUnsavedChanges from "hooks/useWarnIfUnsavedChanges"
 import { useRef } from "react"
-import { FormProvider, useForm } from "react-hook-form"
-import { RoleFormType, Visibility } from "types"
-import getRandomInt from "utils/getRandomInt"
-import SetVisibility from "../../SetVisibility"
+import { FormProvider } from "react-hook-form"
+import useAddRoleForm from "../hooks/useAddRoleForm"
+import useSubmitAddRole from "../hooks/useSubmitAddRole"
+import AddRoleDrawerHeader from "./AddRoleDrawerHeader"
 
 const AddRoleDrawer = ({ isOpen, onClose, finalFocusRef }): JSX.Element => {
-  const { id } = useGuild()
-
-  const toast = useToast()
-
-  const { onSubmit, isLoading, isSigning, signLoadingText } = useCreateRole({
-    onSuccess: () => {
-      toast({
-        title: "Role successfully created",
-        status: "success",
-      })
-      onClose()
-      methods.reset(defaultValues)
-    },
-  })
-
-  const defaultValues: RoleToCreate = {
-    guildId: id,
-    name: "",
-    description: "",
-    logic: "AND",
-    requirements: [
-      {
-        type: "FREE",
-      },
-    ],
-    roleType: "NEW",
-    imageUrl: `/guildLogos/${getRandomInt(286)}.svg`,
-    visibility: Visibility.PUBLIC,
-    rolePlatforms: [],
-  }
-
-  const methods = useForm<RoleFormType>({
-    mode: "all",
-    defaultValues,
-  })
+  const methods = useAddRoleForm()
 
   /**
    * TODO: for some reason, isDirty was true & dirtyFields was an empty object and I
@@ -87,33 +44,18 @@ const AddRoleDrawer = ({ isOpen, onClose, finalFocusRef }): JSX.Element => {
   } = useDisclosure()
 
   const onCloseAndClear = () => {
-    methods.reset(defaultValues)
+    methods.reset(methods.defaultValues)
     onAlertClose()
     onClose()
   }
 
-  const iconUploader = usePinata({
-    fieldToSetOnSuccess: "imageUrl",
-    fieldToSetOnError: "imageUrl",
-    control: methods.control,
-  })
-
   const drawerBodyRef = useRef<HTMLDivElement>()
 
-  const { handleSubmit, isUploadingShown, uploadLoadingText } = useSubmitWithUpload(
-    methods.handleSubmit(onSubmit, (formErrors) => {
-      if (formErrors.requirements && drawerBodyRef.current) {
-        drawerBodyRef.current.scrollBy({
-          top: drawerBodyRef.current.scrollHeight,
-          behavior: "smooth",
-        })
-      }
-    }),
-    iconUploader.isUploading
-  )
-
-  const loadingText = signLoadingText || uploadLoadingText || "Saving data"
-  const setVisibilityModalProps = useVisibilityModalProps()
+  const { onSubmit, isLoading, loadingText, iconUploader } = useSubmitAddRole({
+    drawerBodyRef,
+    methods,
+    onSuccess: onClose,
+  })
 
   return (
     <>
@@ -128,29 +70,7 @@ const AddRoleDrawer = ({ isOpen, onClose, finalFocusRef }): JSX.Element => {
         <DrawerContent>
           <DrawerBody ref={drawerBodyRef} className="custom-scrollbar">
             <FormProvider {...methods}>
-              <DrawerHeader
-                title="Add role"
-                justifyContent="start"
-                spacing={1}
-                alignItems="center"
-              >
-                <Box>
-                  <SetVisibility
-                    entityType="role"
-                    defaultValues={{
-                      visibility: methods.getValues("visibility"),
-                      visibilityRoleId: methods.getValues("visibilityRoleId"),
-                    }}
-                    onSave={({ visibility, visibilityRoleId }) => {
-                      methods.setValue("visibility", visibility)
-                      methods.setValue("visibilityRoleId", visibilityRoleId)
-                      setVisibilityModalProps.onClose()
-                    }}
-                    {...setVisibilityModalProps}
-                  />
-                </Box>
-              </DrawerHeader>
-
+              <AddRoleDrawerHeader />
               <VStack spacing={10} alignItems="start">
                 <AddRolePlatforms />
 
@@ -175,10 +95,10 @@ const AddRoleDrawer = ({ isOpen, onClose, finalFocusRef }): JSX.Element => {
               Cancel
             </Button>
             <Button
-              isLoading={isLoading || isSigning || isUploadingShown}
+              isLoading={isLoading}
               colorScheme="green"
               loadingText={loadingText}
-              onClick={handleSubmit}
+              onClick={onSubmit}
               data-test="save-role-button"
             >
               Save
